src/requirements.txt
urllib3==1.25.11
schedule==1.0.0
pytz==2020.5
influxdb_client==1.14.0
httpx==0.16.1
pickleDB==0.9.2
PyYAML==5.4.1
